fix: cors middleware created, eslint err fixed
