+++ middlewares/cors.js
@@ -1,0 +1,25 @@
+const allowedCors = [
+  'http://localhost:3000',
+  'https://m.explorer.nomoredomains.sbs',
+  'http://m.explorer.nomoredomains.sbs',
+];
+
+module.exports = (req, res, next) => {
+  const { method } = req;
+  const { origin } = req.headers;
+  const requestHeaders = req.headers['access-control-request-headers'];
+  const DEFAULT_ALLOWED_METHODS = 'GET,HEAD,PUT,PATCH,POST,DELETE';
+
+  if (allowedCors.includes(origin)) {
+    res.header('Access-Control-Allow-Origin', origin);
+    res.header('Access-Control-Allow-Credentials', true);
+  }
+
+  if (method === 'OPTIONS') {
+    res.header('Access-Control-Allow-Methods', DEFAULT_ALLOWED_METHODS);
+    res.header('Access-Control-Allow-Headers', requestHeaders);
+    return res.end();
+  }
+
+  return next();
+};
